requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
deprecation==2.1.0
dj-database-url==3.0.1
Django==5.2.7
django-allauth==65.12.1
gotrue==2.12.4
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
invoke==2.2.1
multidict==6.7.0
oauthlib==3.3.1
packaging==25.0
paramiko==4.0.0
pillow==12.0.0
postgrest==2.22.4
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
realtime==2.22.4
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage==0.0.4.3
storage3==2.22.4
StrEnum==0.4.15
stripe==13.1.0
supabase==2.22.4
supabase-auth==2.22.4
supabase-functions==2.22.4
supafunc==0.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
whitenoise==6.11.0
yarl==1.22.0
